+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - `jxl-color`: Use better PQ to HLG method (#348).
 
+### Fixed
+- `jxl-render`: Fix requested color encoding not applied in some cases (#369).
+
 ## [0.9.1] - 2024-10-12
 
 ### Fixed