+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.10.0] - 2024-11-01
 
 ### Added
 - `jxl-oxide`: Accept `u8` and `u16` output buffers in `ImageStream::write_to_buffer` (#366).
@@ -148,7 +148,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - This is the first official release of jxl-oxide, a JPEG XL decoder written in Rust.
 - Frequently used features are mostly implemented.
 
-[Unreleased]: https://github.com/tirr-c/jxl-oxide/compare/0.9.1...HEAD
+[0.10.0]: https://github.com/tirr-c/jxl-oxide/releases/tag/0.10.0
 [0.9.1]: https://github.com/tirr-c/jxl-oxide/releases/tag/0.9.1
 [0.9.0]: https://github.com/tirr-c/jxl-oxide/releases/tag/0.9.0
 [0.8.1]: https://github.com/tirr-c/jxl-oxide/releases/tag/0.8.1